Initial Centralized Code – Version 1.0
This first version of the centralized code sets up an integrated web server on the ESP32 that collects sensor data and makes it viewable in real time over a common internet connection.

Features:
Real-Time Sensor Data Collection, Unified Data Packaging, Built-In Web Server and Modular and Scalable Design for future edition

This will potentially serve as a test bench for our sensor data, demonstrating the ability to centralize and visualize data in real time.